libs/button/src/lib/button-link/button-link.component.ts
import {
  ChangeDetectionStrategy,
  Component,
  Input,
  ViewEncapsulation,
} from '@angular/core';

import { ButtonCore } from '../button-core/button-core';

@Component({
  selector: 'spy-button-link',
  templateUrl: './button-link.component.html',
  styleUrls: ['./button-link.component.less'],
  changeDetection: ChangeDetectionStrategy.OnPush,
  encapsulation: ViewEncapsulation.None,
})
export class ButtonLinkComponent extends ButtonCore {
  @Input() url?: string;

  protected buttonClassName = 'spy-button-link';
}
